Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    int r;
-    scanf("%d", &r);
-    int area= (3.14*r*r);
-    printf("Area: %d\n", area);
+    float r;
+    scanf("%f", &r);
+    float area= (3.14*r*r);
+    printf("Area: %.2f\n", area);
     return 0;
 }
